split preference (Kotlin)
