+++ vector/src/main/res/xml/vector_settings_flair.xml
@@ -1,12 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.preference.PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+<androidx.preference.PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
 
     <im.vector.riotredesign.core.preference.VectorPreferenceCategory
         android:key="SETTINGS_GROUPS_FLAIR_KEY"
         android:title="@string/settings_flair" />
-
-    <im.vector.riotredesign.core.preference.VectorPreferenceDivider />
 
 </androidx.preference.PreferenceScreen>
