+++ vector/src/main/res/xml/vector_settings_preferences.xml
