+++ vector/src/main/res/xml/vector_settings_root.xml
@@ -6,13 +6,13 @@
         android:layout_width="match_parent"
         android:icon="@drawable/ic_settings_root_general"
         android:title="@string/settings_general_title"
-        app:fragment="com.example.SyncFragment" />
+        app:fragment="im.vector.riotredesign.features.settings.VectorSettingsGeneral" />
 
     <im.vector.riotredesign.core.preference.VectorPreference
         android:layout_width="match_parent"
         android:icon="@drawable/ic_settings_root_flair"
         android:title="@string/settings_flair"
-        app:fragment="com.example.SyncFragment" />
+        app:fragment="im.vector.riotredesign.features.settings.VectorSettingsFlair" />
 
     <im.vector.riotredesign.core.preference.VectorPreference
         android:layout_width="match_parent"
@@ -25,43 +25,37 @@
         android:layout_width="match_parent"
         android:icon="@drawable/ic_settings_root_preferences"
         android:title="@string/settings_preferences"
-        app:fragment="com.example.SyncFragment" />
+        app:fragment="im.vector.riotredesign.features.settings.VectorSettingsPreferences" />
 
     <im.vector.riotredesign.core.preference.VectorPreference
         android:layout_width="match_parent"
         android:icon="@drawable/ic_settings_root_call"
         android:title="@string/preference_voice_and_video"
-        app:fragment="com.example.SyncFragment" />
+        app:fragment="im.vector.riotredesign.features.settings.VectorSettingsVoiceVideo" />
 
-    <!-- TODO Ignored users -->
+    <!-- TODO Ignored users icon -->
     <im.vector.riotredesign.core.preference.VectorPreference
         android:layout_width="match_parent"
-        android:icon="@drawable/ic_settings_root_security_privacy"
+        android:icon="@drawable/ic_settings_root_legacy"
         android:title="@string/settings_ignored_users"
-        app:fragment="com.example.SyncFragment" />
+        app:fragment="im.vector.riotredesign.features.settings.VectorSettingsIgnoredUsers" />
 
     <im.vector.riotredesign.core.preference.VectorPreference
         android:layout_width="match_parent"
         android:icon="@drawable/ic_settings_root_security_privacy"
         android:title="@string/settings_security_and_privacy"
-        app:fragment="com.example.SyncFragment" />
+        app:fragment="im.vector.riotredesign.features.settings.VectorSettingsSecurityPrivacy" />
 
     <im.vector.riotredesign.core.preference.VectorPreference
         android:layout_width="match_parent"
         android:icon="@drawable/ic_settings_root_labs"
         android:title="@string/room_settings_labs_pref_title"
-        app:fragment="com.example.SyncFragment" />
-
-    <im.vector.riotredesign.core.preference.VectorPreference
-        android:layout_width="match_parent"
-        android:icon="@drawable/ic_settings_root_legacy"
-        android:title="@string/preference_root_legacy"
-        app:fragment="com.example.SyncFragment" />
+        app:fragment="im.vector.riotredesign.features.settings.VectorSettingsLabs" />
 
     <im.vector.riotredesign.core.preference.VectorPreference
         android:layout_width="match_parent"
         android:icon="@drawable/ic_settings_root_help_about"
         android:title="@string/preference_root_help_about"
-        app:fragment="com.example.SyncFragment" />
+        app:fragment="im.vector.riotredesign.features.settings.VectorSettingsHelpAbout" />
 
 </androidx.preference.PreferenceScreen>
