+++ vector/src/main/res/xml/vector_settings_security_privacy.xml
@@ -72,6 +72,4 @@
             android:title="@string/send_bug_report_rage_shake" />
     </im.vector.riotredesign.core.preference.VectorPreferenceCategory>
 
-    <im.vector.riotredesign.core.preference.VectorPreferenceDivider />
-
 </androidx.preference.PreferenceScreen>
